Test Acc: Remove flatmap
This commit removes the flatmap package from tests since it has been
deprecated.
