+++ openstack/util.go
@@ -3,7 +3,6 @@ package openstack
 import (
 	"fmt"
 	"net/http"
-	"reflect"
 	"sort"
 	"strconv"
 	"strings"
@@ -11,10 +10,8 @@ import (
 
 	"github.com/Unknwon/com"
 	"github.com/gophercloud/gophercloud"
-	"github.com/hashicorp/terraform/flatmap"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/hashicorp/terraform/terraform"
 )
 
 // BuildRequest takes an opts struct and builds a request body for
@@ -189,40 +186,6 @@ func networkV2AttributesTags(d *schema.ResourceData) (tags []string) {
 		tags[i] = raw.(string)
 	}
 	return
-}
-
-func testAccCheckNetworkingV2Tags(name string, tags []string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		rs, ok := s.RootModule().Resources[name]
-
-		if !ok {
-			return fmt.Errorf("resource not found: %s", name)
-		}
-
-		var tagLen int64
-		var err error
-		if count, ok := rs.Primary.Attributes["tags.#"]; !ok {
-			return fmt.Errorf("resource tags not found: %s.tags", name)
-		} else {
-			tagLen, err = strconv.ParseInt(count, 10, 64)
-			if err != nil {
-				return fmt.Errorf("Failed to parse tag amount: %s", err)
-			}
-		}
-
-		rtags := make([]string, tagLen)
-		itags := flatmap.Expand(rs.Primary.Attributes, "tags").([]interface{})
-		for i, val := range itags {
-			rtags[i] = val.(string)
-		}
-		sort.Strings(rtags)
-		sort.Strings(tags)
-		if !reflect.DeepEqual(rtags, tags) {
-			return fmt.Errorf(
-				"%s.tags: expected: %#v, got %#v", name, tags, rtags)
-		}
-		return nil
-	}
 }
 
 func expandToMapStringString(v map[string]interface{}) map[string]string {
